services/budobserve/budobserve/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "budobserve"
version = "0.1.0"
description = "Observability SDK for Bud-Stack platform - built on OpenTelemetry"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Bud Ecosystem", email = "dev@budecosystem.com" }
]
keywords = [
    "observability",
    "opentelemetry",
    "tracing",
    "metrics",
    "logging",
    "llm",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
dependencies = [
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-instrumentation>=0.49b0",
    "typing-extensions>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "pyright>=1.1.380",
    "mypy>=1.11.0",
]
fastapi = [
    "opentelemetry-instrumentation-fastapi>=0.49b0",
]
httpx = [
    "opentelemetry-instrumentation-httpx>=0.49b0",
]
sqlalchemy = [
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
]
redis = [
    "opentelemetry-instrumentation-redis>=0.49b0",
]
all = [
    "budobserve[fastapi,httpx,sqlalchemy,redis]",
]

[project.urls]
Homepage = "https://github.com/BudEcosystem/bud-runtime"
Documentation = "https://github.com/BudEcosystem/bud-runtime/tree/master/services/budobserve"
Repository = "https://github.com/BudEcosystem/bud-runtime"

[tool.hatch.build.targets.wheel]
packages = ["budobserve"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "D",      # pydocstyle
    "RUF",    # Ruff-specific rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["budobserve"]

[tool.pyright]
include = ["budobserve"]
exclude = ["**/__pycache__"]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
